src/requirements.txt
numpy==1.24.3
pydantic==1.10.7
pyyaml==6.0
pyspark==3.4.0
tqdm==4.65.0
matplotlib
seaborn
pandas
